requirements.txt
discord.py
python-dotenv
aiohttp
logging
